Minor changes to variables
I had to go through and just make sure the variables are set to SENSIBLE defaults, as I did incorporate a lot of new features into this script. One of these features is the capability to kill init with SIGKILL, should the koops variable be set to 1 and the script run into an actual exit code it should NEVER, EVER see in a billion years unless it's running on an actual corrupted system, or is running on hardware breathing its last breath of air through its coolers.