Script overhaul (#1)
* Reflect new changes in script

This update to the README reflects and shows some new changes made in the shellscript (*sh form) version of this script. It is designed with all of the most popular *sh variants in mind; the only one I could NOT get working with generic POSIX compliant code was FISH (I'm sorry for you older Gentoo install users! I wish I knew what I was doing with FISH!)
However with nearly everyone having DASH, ASH, BASH, ZSH, or KSH at this point, I'm pretty sure it's safe to assume just about anyone on Linux can run this script with zero problems.

* Delete bash_start.sh

* Upload overhauled script

Uploading the new, vastly improved, and much tweaked script containing very useful functionality.

* Minor changes to variables

I had to go through and just make sure the variables are set to SENSIBLE defaults, as I did incorporate a lot of new features into this script. One of these features is the capability to kill init with SIGKILL, should the koops variable be set to 1 and the script run into an actual exit code it should NEVER, EVER see in a billion years unless it's running on an actual corrupted system, or is running on hardware breathing its last breath of air through its coolers.

* Fix README formatting