Update shellscript version
Corrected some possible problems, expanded download mode, as well as expanded location auto-detect. You can now force the script to start a fresh instance in your specified path as long as Download Mode is enabled (it isn't by default), or if you so choose, you can also have it handle the entire shebang from run #1 in its current directory.

More features to come as I figure out ways to improve the script's usability and robustness. As of right now, I don't have a way to check for WRITE access to a specified directory beyond just trying to create a .running file where the server is located to indicate that the server is, in fact, running, or had previously run into an abrupt crash preventing its removal. I should have write access detection earlier in the script to be robust, but the current check is simple enough, as long as the directory doesn't have completely broken permissions in the first place.