Update start.sh (#6)
Fixed a bug that would disable download mode completely upon the script finding the current and new versions match. This was not intentional behavior; I simply wanted to have it not commit to the download if the versions matched. However, I have found it is difficult to keep track of the old and new versions, and the way I have implemented version control is rather crude.