Giantvince1 patch 1 (#12)
* Replacing some files.

Just replacing the scripts with some actual fixed versions. A few things were not quite right, so I had to dig in and fix them up a bit.

* Update stop.sh

* Get rid of interactive variable from exclusively interactive version

* Update start_interact.sh

Add in final running check that was missing in interactive-exclusive

* Update start.sh

Fix this script alongside the interactive-exclusive version to patch in the final running check that was missing.